LayoutTests/media/modern-media-controls/tracks-support/ios/tracks-support-click-track-in-contextmenu.html

<!DOCTYPE html>
<script src="../../resources/media-controls-utils.js"></script>
<script src="../../../../resources/ui-helper.js"></script>
<script src="../../../../resources/js-test-pre.js"></script>
<body>
<video src="../../../content/audio-tracks.mp4" style="position: absolute; left: 0; top: 0; width: 300px;" controls autoplay muted playsinline></video>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Clicking on a track in the tracks contextmenu.");

const media = document.querySelector("video");
const shadowRoot = window.internals.shadowRoot(media);

let previouslySelectedTrack = 0;

media.addEventListener("play", () => {
    media.pause();

    dumpAudioTracksEnabledState();

    media.audioTracks.addEventListener("change", handleAudioTracksChange);

    shouldBecomeDifferent("shadowRoot.querySelector('button.tracks')", "null", () => {
        shouldBecomeDifferent("shadowRoot.querySelector('button.tracks').getBoundingClientRect().width", "0", () => {
            clickOnTracksButton(clickOnSecondAudioTrack);
        });
    });
});

function clickOnTracksButton(continuation)
{
    debug("Tapping tracks button...");
    pressOnElement(shadowRoot.querySelector("button.tracks"), () => {
        window.requestAnimationFrame(continuation);
    });
}

async function clickOnSecondAudioTrack()
{
    debug("Selecting second audio track...");
    let contextmenu = await getTracksContextMenu();
    let languagesMenu = contextmenu[0].children[0];
    await UIHelper.chooseMenuAction(languagesMenu.title);
    await UIHelper.delayFor(500); // give time for the UIMenu to animate
    await UIHelper.chooseMenuAction(languagesMenu.children[1].title);
}

async function clickOnThirdAudioTrack()
{
    debug("Selecting third audio track...");
    let contextmenu = await getTracksContextMenu();
    let languagesMenu = contextmenu[0].children[0];
    await UIHelper.chooseMenuAction(languagesMenu.title);
    await UIHelper.delayFor(500); // give time for the UIMenu to animate
    await UIHelper.chooseMenuAction(languagesMenu.children[2].title);
}

async function handleAudioTracksChange()
{
    const newSelectedTrack = Array.from(media.audioTracks).findIndex(track => track.enabled);
    if (previouslySelectedTrack === newSelectedTrack)
        return;

    previouslySelectedTrack = newSelectedTrack;

    debug("Audio tracks changed.");

    dumpAudioTracksEnabledState();

    await UIHelper.waitForContextMenuToHide();

    if (media.audioTracks[1].enabled)
        clickOnTracksButton(clickOnThirdAudioTrack);
    else if (media.audioTracks[2].enabled) {
        media.remove();
        finishJSTest();
    }
}

function dumpAudioTracksEnabledState()
{
    Array.from(media.audioTracks).forEach((audioTrack, index) => {
        debug(`media.audioTracks[${index}].enabled = ${media.audioTracks[index].enabled}`);
    });
}

</script>
<script src="../../../../resources/js-test-post.js"></script>
</body>
