LayoutTests/media/modern-media-controls/tracks-support/ios/tracks-support-click-track-in-contextmenu-expected.txt

Clicking on a track in the tracks contextmenu.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


media.audioTracks[0].enabled = true
media.audioTracks[1].enabled = false
media.audioTracks[2].enabled = false
PASS shadowRoot.querySelector('button.tracks') became different from null
PASS shadowRoot.querySelector('button.tracks').getBoundingClientRect().width became different from 0
Tapping tracks button...
Selecting second audio track...
Audio tracks changed.
media.audioTracks[0].enabled = false
media.audioTracks[1].enabled = true
media.audioTracks[2].enabled = false
Tapping tracks button...
Selecting third audio track...
Audio tracks changed.
media.audioTracks[0].enabled = false
media.audioTracks[1].enabled = false
media.audioTracks[2].enabled = true
PASS successfullyParsed is true

TEST COMPLETE
